+++ src/renderer/SideMenu.tsx
@@ -39,6 +39,7 @@ import { ScrollArea } from './components/ScrollArea/ScrollArea';
 import UpdateNotifier from './containers/UpdateNotifier/UpdateNotifier';
 import CloudStatusCard from './containers/ApplicationStatusCard/CloudStatusCard';
 import { Phrase } from 'localisation/phrases';
+import PatreonButton from './PatreonButton';
 
 interface IProps {
   recorderStatus: RecStatus;
@@ -228,6 +229,7 @@ const SideMenu = (props: IProps) => {
           <LogsButton appState={appState} />
           <TestButton recorderStatus={recorderStatus} appState={appState} />
           <DiscordButton appState={appState} />
+          <PatreonButton appState={appState} />
         </div>
         {!!appVersion && (
           <div className="w-full mt-1 text-foreground font-sans text-[11px] font-bold text-center opacity-75">