+++ src/renderer/components/Tables/Cells.tsx
@@ -263,8 +263,7 @@ export const populateAffixesCell = (
 };
 
 export const populateViewpointCell = (
-  info: CellContext<RendererVideo, unknown>,
-  preferredViewpoint: string,
+  info: CellContext<RendererVideo, unknown>
 ) => {
   const video = info.getValue() as RendererVideo;
   const count = countUniqueViewpoints(video);
@@ -272,18 +271,7 @@ export const populateViewpointCell = (
   // Prioritize the any videos with a disk copy as that's likely to be the
   // local users viewpoint so most relevant to them.
   const povs = [video, ...video.multiPov].sort(povDiskFirstNameSort);
-  let first = povs[0];
-
-  if (preferredViewpoint) {
-    // If we have a preferred viewpoint, try to put that player first.
-    const preferredMatch = povs.find(
-      (p) => p.player?._name === preferredViewpoint,
-    );
-
-    if (preferredMatch) {
-      first = preferredMatch;
-    }
-  }
+  const first = povs[0];
 
   const { player } = first;