+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - [PR 768](https://github.com/aza547/wow-recorder/pull/768) - Added config for setting classic PTR. 
+- [Issue 776](https://github.com/aza547/wow-recorder/issues/776) - Patreon icon in bottom left.
 
 ### Fixed
 - Improve frontend render performance on upload/download progress re-renders.